🏗 Refactored: – Refactored types and Swagger decorators

+++ src/authors/authors.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { AuthorsEntity } from './authors.entity';
@@ -12,6 +12,10 @@ export class AuthorsService {
   ) {}
 
   async create(dto: AuthorsDto): Promise<AuthorsEntity> {
+    const twin = await this.authorsRepository.findOne({
+      where: { surname: dto.surname },
+    });
+    if (twin) throw new HttpException('Author already exists', 409);
     const author = this.authorsRepository.create(dto);
     return await this.authorsRepository.save(author);
   }
@@ -35,12 +39,20 @@ export class AuthorsService {
   }
 
   async update(id: number, dto: AuthorsDto): Promise<AuthorsEntity> {
-    await this.authorsRepository.update(id, dto);
-    const author = await this.authorsRepository.findOne({ where: { id } });
-    return author;
+    const author = await this.getById(id);
+
+    if (!(await this.getById(id)))
+      throw new HttpException('Book not found', 404);
+
+    const modified = await this.authorsRepository.merge(author, dto);
+    await this.authorsRepository.save(modified);
+
+    return await this.authorsRepository.findOne({ where: { id } });
   }
 
   async delete(id: number): Promise<void> {
+    const author = await this.getById(id);
+    if (!author) throw new HttpException('Author not found', 404);
     await this.authorsRepository.delete(id);
   }
 }
